requirements.txt
comm==0.1.3
dash==2.10.2
ipython==8.14.0
ipywidgets==8.0.6
jupyter_dash==0.4.2
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.7.0.72
opencv_python_headless==4.5.5.64
pandas==2.0.2
Pillow==9.5.0
plotly==5.15.0
salesforce_lavis==1.0.2
scikit_learn==1.2.2
torch==2.0.1
transformers==4.26.1
wasabi==1.1.1
